akencoder: impl eval process for producer thread

+++ akashi_engine/src/libakencoder/encode_queue.h
@@ -47,7 +47,7 @@ namespace akashi {
     namespace encoder {
 
         struct EncodeQueueData {
-            int test_data = -1;
+            double test_data = -1;
         };
 
         class EncodeQueue final {
